Added VUser (alternate User, save for -enroll)
Haven't done -enroll yet, but the rest of it is up. You can see a lot of
the strengths of this style of implementation here. For example: the
definition for unenroll in line 49 will now only take a student without
us having to test for it. I added functionality there to test who is
doing the removal...mostly to show off right now.

As an aside, maybe some exception handeling if we're passed the wrong
type would be helpful anyway. Although I think we have enough control
over who can call this that that shouldn't be an issue.